Move lattice reduction interface to its own submodule
