+++ docs/api/autodoc/spglib.reduce.rst
@@ -1,0 +1,8 @@
+spglib.reduce module
+====================
+
+.. automodule:: spglib.reduce
+   :members:
+   :private-members:
+   :show-inheritance:
+   :undoc-members:
